- Introduce Dynamic proxy for auto reconnect decorator. - Fix autoreconnection for opc ua - Add unit test cases
